Files from plain implementation based on the Truffle pet shop example
The truffle-contract and Web3 versions are very old and do not support the latest Solidity functions, most notably returning structs. I tried updating Web3 to the latest 1.x, this threw errors so I tried updating truffle-contract, which is packaged in a kind of weird way with the example (just a .js file). Then tried installing `@truffle/contract` npm package, not working yet but that's the direction I'm going.